Rename prover backends module
