+++ crates/l2/prover/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod backends;
+pub mod backend;
 pub mod prover;
 
 pub mod config;
@@ -11,10 +11,10 @@ pub async fn init_client(config: ProverConfig) {
 }
 
 #[cfg(feature = "risc0")]
-pub use backends::risc0::*;
+pub use backend::risc0::*;
 
 #[cfg(feature = "sp1")]
-pub use backends::sp1::*;
+pub use backend::sp1::*;
 
 #[cfg(not(any(feature = "risc0", feature = "sp1")))]
-pub use backends::exec::*;
+pub use backend::exec::*;
